+++ src/formatters/luau.rs
@@ -8,13 +8,14 @@ use crate::{
             format_punctuated_multiline, format_symbol, format_token_reference,
             try_format_punctuated, EndTokenType,
         },
-        table::{create_table_braces, TableType},
+        table::{create_table_braces, format_multiline_table, format_singleline_table, TableType},
         trivia::{
-            strip_leading_trivia, strip_trailing_trivia, FormatTriviaType, UpdateLeadingTrivia,
-            UpdateTrailingTrivia,
+            strip_leading_trivia, strip_trailing_trivia, strip_trivia, FormatTriviaType,
+            UpdateLeadingTrivia, UpdateTrailingTrivia,
         },
         trivia_util::{
-            contains_comments, take_type_field_trailing_comments, token_trivia_contains_comments,
+            contains_comments, token_trivia_contains_comments, trivia_is_comment,
+            type_info_trailing_trivia,
         },
     },
     shape::Shape,
@@ -23,10 +24,7 @@ use full_moon::ast::types::{
     CompoundAssignment, CompoundOp, ExportedTypeDeclaration, GenericDeclaration, IndexedTypeInfo,
     TypeArgument, TypeAssertion, TypeDeclaration, TypeField, TypeFieldKey, TypeInfo, TypeSpecifier,
 };
-use full_moon::ast::{
-    punctuated::{Pair, Punctuated},
-    span::ContainedSpan,
-};
+use full_moon::ast::{punctuated::Punctuated, span::ContainedSpan};
 use full_moon::tokenizer::{Token, TokenReference, TokenType};
 use std::boxed::Box;
 
@@ -71,7 +69,7 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
                 fmt_symbol!(ctx, start_brace, "{ ", shape),
                 fmt_symbol!(ctx, end_brace, " }", shape),
             );
-            let type_info = Box::new(format_type_info(ctx, type_info, shape));
+            let type_info = Box::new(format_type_info(ctx, type_info, shape + 2)); // 2 = "{ "
 
             TypeInfo::Array { braces, type_info }
         }
@@ -91,7 +89,14 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
 
             let force_multiline = token_trivia_contains_comments(start_parens.trailing_trivia())
                 || token_trivia_contains_comments(end_parens.leading_trivia())
-                || contains_comments(arguments);
+                || contains_comments(arguments)
+                || shape
+                    .add_width(
+                        2 + 4
+                            + arguments.to_string().len()
+                            + strip_trailing_trivia(&**return_type).to_string().len(),
+                    )
+                    .over_budget(); // 2 = opening/closing parens, 4 = " -> "
 
             let (parentheses, arguments, shape) = if force_multiline {
                 let start_parens = fmt_symbol!(ctx, start_parens, "(", shape)
@@ -145,7 +150,13 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
         } => {
             let base = format_token_reference(ctx, base, shape);
             let arrows = format_contained_span(ctx, arrows, shape);
-            let generics = try_format_punctuated(ctx, generics, shape, format_type_info, None);
+            let generics = try_format_punctuated(
+                ctx,
+                generics,
+                shape + (strip_trivia(&base).to_string().len() + 1), // 1 = "<"
+                format_type_info,
+                None,
+            );
             TypeInfo::Generic {
                 base,
                 arrows,
@@ -160,7 +171,7 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
         } => {
             let left = Box::new(format_type_info(ctx, left, shape));
             let ampersand = fmt_symbol!(ctx, ampersand, " & ", shape);
-            let right = Box::new(format_type_info(ctx, right, shape));
+            let right = Box::new(format_type_info(ctx, right, shape + 3)); // 3 = " & "
             TypeInfo::Intersection {
                 left,
                 ampersand,
@@ -175,7 +186,11 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
         } => {
             let module = format_token_reference(ctx, module, shape);
             let punctuation = fmt_symbol!(ctx, punctuation, ".", shape);
-            let type_info = Box::new(format_indexed_type_info(ctx, type_info, shape));
+            let type_info = Box::new(format_indexed_type_info(
+                ctx,
+                type_info,
+                shape + (strip_trivia(&module).to_string().len() + 1), // 1 = "."
+            ));
             TypeInfo::Module {
                 module,
                 punctuation,
@@ -197,75 +212,46 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
 
         TypeInfo::Table { braces, fields } => {
             let (start_brace, end_brace) = braces.tokens().to_owned();
-            let braces_range = (
-                start_brace.token().end_position().bytes(),
-                end_brace.token().start_position().bytes(),
-            );
+            let contains_comments = start_brace.trailing_trivia().any(trivia_is_comment)
+                || end_brace.leading_trivia().any(trivia_is_comment)
+                || fields.pairs().any(|field| {
+                    contains_comments(field.punctuation()) || contains_comments(field.value())
+                });
 
-            let mut current_fields = fields.to_owned().into_pairs().peekable();
-            let is_multiline = (braces_range.1 - braces_range.0) > 30; // TODO: Properly determine this arbitrary number, and see if other factors should come into play
-            let table_type = match current_fields.peek() {
-                Some(_) => match is_multiline {
-                    true => TableType::MultiLine,
-                    false => TableType::SingleLine,
-                },
-                None => TableType::Empty,
-            };
+            let table_type = match (contains_comments, fields.iter().next()) {
+                // Table contains comments, so force multiline
+                (true, _) => TableType::MultiLine,
 
-            let braces = create_table_braces(ctx, start_brace, end_brace, table_type, shape);
+                (false, Some(_)) => {
+                    let braces_range = (
+                        start_brace.token().end_position().bytes(),
+                        end_brace.token().start_position().bytes(),
+                    );
 
-            let shape = if is_multiline {
-                shape.increment_additional_indent()
-            } else {
-                shape
-            };
+                    let singleline_shape = shape + (braces_range.1 - braces_range.0);
 
-            let mut fields = Punctuated::new();
-
-            while let Some(pair) = current_fields.next() {
-                let (field, punctuation) = pair.into_tuple();
-
-                let leading_trivia = match is_multiline {
-                    true => FormatTriviaType::Append(vec![create_indent_trivia(ctx, shape)]),
-                    false => FormatTriviaType::NoChange,
-                };
-
-                let mut formatted_field = format_type_field(ctx, &field, leading_trivia, shape);
-                let mut formatted_punctuation = None;
-
-                match is_multiline {
-                    true => {
-                        // Continue adding a comma and a new line for multiline tables
-                        // Add newline trivia to the end of the symbol
-
-                        let (field, mut trailing_comments) =
-                            take_type_field_trailing_comments(formatted_field);
-                        formatted_field = field;
-                        trailing_comments.push(create_newline_trivia(ctx));
-
-                        let symbol = match punctuation {
-                            Some(punctuation) => fmt_symbol!(ctx, &punctuation, ",", shape),
-                            None => TokenReference::symbol(",").unwrap(),
-                        }
-                        .update_trailing_trivia(FormatTriviaType::Append(trailing_comments));
-                        formatted_punctuation = Some(symbol)
-                    }
-
-                    false => {
-                        if current_fields.peek().is_some() {
-                            // Have more elements still to go
-                            formatted_punctuation = match punctuation {
-                                Some(punctuation) => {
-                                    Some(fmt_symbol!(ctx, &punctuation, ", ", shape))
-                                }
-                                None => Some(TokenReference::symbol(", ").unwrap()),
-                            }
-                        };
+                    match singleline_shape.over_budget() {
+                        true => TableType::MultiLine,
+                        false => TableType::SingleLine,
                     }
                 }
 
-                fields.push(Pair::new(formatted_field, formatted_punctuation));
-            }
+                (false, None) => TableType::Empty,
+            };
+
+            let (braces, fields) = match table_type {
+                TableType::Empty => {
+                    let braces =
+                        create_table_braces(ctx, start_brace, end_brace, table_type, shape);
+                    (braces, Punctuated::new())
+                }
+                TableType::SingleLine => {
+                    format_singleline_table(ctx, braces, fields, format_type_field, shape)
+                }
+                TableType::MultiLine => {
+                    format_multiline_table(ctx, braces, fields, format_type_field, shape)
+                }
+            };
 
             TypeInfo::Table { braces, fields }
         }
@@ -287,8 +273,9 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
                 ),
                 shape,
             );
+            let shape = shape + 6; // 6 = "typeof"
             let parentheses = format_contained_span(ctx, parentheses, shape);
-            let inner = Box::new(format_expression(ctx, inner, shape));
+            let inner = Box::new(format_expression(ctx, inner, shape + 1)); // 1 = "("
             TypeInfo::Typeof {
                 typeof_token,
                 parentheses,
@@ -298,7 +285,7 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
 
         TypeInfo::Tuple { parentheses, types } => {
             let parentheses = format_contained_span(ctx, parentheses, shape);
-            let types = try_format_punctuated(ctx, types, shape, format_type_info, None);
+            let types = try_format_punctuated(ctx, types, shape + 1, format_type_info, None); // 1 = "("
 
             TypeInfo::Tuple { parentheses, types }
         }
@@ -306,14 +293,14 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
         TypeInfo::Union { left, pipe, right } => {
             let left = Box::new(format_type_info(ctx, left, shape));
             let pipe = fmt_symbol!(ctx, pipe, " | ", shape);
-            let right = Box::new(format_type_info(ctx, right, shape));
+            let right = Box::new(format_type_info(ctx, right, shape + 3)); // 3 = " | "
 
             TypeInfo::Union { left, pipe, right }
         }
 
         TypeInfo::Variadic { ellipse, type_info } => {
             let ellipse = fmt_symbol!(ctx, ellipse, "...", shape);
-            let type_info = Box::new(format_type_info(ctx, type_info, shape));
+            let type_info = Box::new(format_type_info(ctx, type_info, shape + 3)); // 3 = "..."
 
             TypeInfo::Variadic { ellipse, type_info }
         }
@@ -353,7 +340,14 @@ pub fn format_indexed_type_info(
         } => {
             let base = format_token_reference(ctx, base, shape);
             let arrows = format_contained_span(ctx, arrows, shape);
-            let generics = try_format_punctuated(ctx, generics, shape, format_type_info, None);
+            let generics = try_format_punctuated(
+                ctx,
+                generics,
+                shape + (strip_trivia(&base).to_string().len() + 1), // 1 = "<"
+                format_type_info,
+                None,
+            );
+
             IndexedTypeInfo::Generic {
                 base,
                 arrows,
@@ -376,7 +370,14 @@ fn format_type_argument(ctx: &Context, type_argument: &TypeArgument, shape: Shap
         None => None,
     };
 
-    let type_info = format_type_info(ctx, type_argument.type_info(), shape);
+    let type_info = format_type_info(
+        ctx,
+        type_argument.type_info(),
+        shape
+            + name
+                .as_ref()
+                .map_or(0, |(name, _)| strip_trivia(name).to_string().len() + 2), // 2 = ": "
+    );
 
     type_argument
         .to_owned()
@@ -384,21 +385,39 @@ fn format_type_argument(ctx: &Context, type_argument: &TypeArgument, shape: Shap
         .with_type_info(type_info)
 }
 
+/// Formats a [`TypeField`] present inside of a [`TypeInfo::Table`]
+/// Returns the new [`TypeField`] and any trailing trivia associated with its value (as this may need to later be moved).
+/// If the [`TableType`] provided is [`TableType::MultiLine`] then the trailing trivia from the value will be removed.
 pub fn format_type_field(
     ctx: &Context,
     type_field: &TypeField,
-    leading_trivia: FormatTriviaType,
+    table_type: TableType,
     shape: Shape,
-) -> TypeField {
+) -> (TypeField, Vec<Token>) {
+    let leading_trivia = match table_type {
+        TableType::MultiLine => FormatTriviaType::Append(vec![create_indent_trivia(ctx, shape)]),
+        _ => FormatTriviaType::NoChange,
+    };
+
     let key = format_type_field_key(ctx, type_field.key(), leading_trivia, shape);
     let colon_token = fmt_symbol!(ctx, type_field.colon_token(), ": ", shape);
-    let value = format_type_info(ctx, type_field.value(), shape);
+    let shape = shape + (strip_leading_trivia(&key).to_string().len() + 2);
+    let mut value = format_type_info(ctx, type_field.value(), shape);
 
-    type_field
-        .to_owned()
-        .with_key(key)
-        .with_colon_token(colon_token)
-        .with_value(value)
+    let trailing_trivia = type_info_trailing_trivia(&value);
+
+    if let TableType::MultiLine = table_type {
+        value = value.update_trailing_trivia(FormatTriviaType::Replace(vec![]))
+    }
+
+    (
+        type_field
+            .to_owned()
+            .with_key(key)
+            .with_colon_token(colon_token)
+            .with_value(value),
+        trailing_trivia,
+    )
 }
 
 pub fn format_type_field_key(
@@ -414,7 +433,7 @@ pub fn format_type_field_key(
         TypeFieldKey::IndexSignature { brackets, inner } => TypeFieldKey::IndexSignature {
             brackets: format_contained_span(ctx, brackets, shape)
                 .update_leading_trivia(leading_trivia),
-            inner: format_type_info(ctx, inner, shape),
+            inner: format_type_info(ctx, inner, shape + 1), // 1 = "["
         },
         other => panic!("unknown node {:?}", other),
     }
@@ -426,7 +445,7 @@ pub fn format_type_assertion(
     shape: Shape,
 ) -> TypeAssertion {
     let assertion_op = fmt_symbol!(ctx, type_assertion.assertion_op(), " :: ", shape);
-    let cast_to = format_type_info(ctx, type_assertion.cast_to(), shape);
+    let cast_to = format_type_info(ctx, type_assertion.cast_to(), shape + 4); // 4 = " :: "
 
     TypeAssertion::new(cast_to).with_assertion_op(assertion_op)
 }
@@ -458,21 +477,25 @@ fn format_type_declaration(
         type_token = type_token.update_leading_trivia(FormatTriviaType::Append(leading_trivia))
     }
 
+    let shape = shape + 5; // 5 = "type "
     let type_name = format_token_reference(ctx, type_declaration.type_name(), shape);
+    let shape = shape + type_name.to_string().len();
+
     let generics = type_declaration
         .generics()
         .map(|generics| format_generic_declaration(ctx, generics, shape));
-    let mut equal_token = fmt_symbol!(ctx, type_declaration.equal_token(), " = ", shape);
-    let mut type_definition = format_type_info(ctx, type_declaration.type_definition(), shape)
-        .update_trailing_trivia(FormatTriviaType::Append(trailing_trivia));
 
-    let shape = shape
-        .add_width(
-            5 + type_name.to_string().len()
-                + generics.as_ref().map_or(0, |x| x.to_string().len())
-                + 3,
-        ) // 5 = "type ", "3" = " = "
-        .take_last_line(&strip_trailing_trivia(&type_definition));
+    let shape = match generics {
+        Some(ref generics) => shape.take_last_line(&generics),
+        None => shape,
+    };
+
+    let mut equal_token = fmt_symbol!(ctx, type_declaration.equal_token(), " = ", shape);
+    let mut type_definition =
+        format_type_info(ctx, type_declaration.type_definition(), shape + 3) // 3 = " = "
+            .update_trailing_trivia(FormatTriviaType::Append(trailing_trivia));
+
+    let shape = shape.take_last_line(&strip_trailing_trivia(&type_definition));
 
     if shape.over_budget() {
         let shape = shape.increment_additional_indent();
@@ -562,7 +585,7 @@ pub fn format_type_specifier(
     shape: Shape,
 ) -> TypeSpecifier {
     let punctuation = fmt_symbol!(ctx, type_specifier.punctuation(), ": ", shape);
-    let type_info = format_type_info(ctx, type_specifier.type_info(), shape);
+    let type_info = format_type_info(ctx, type_specifier.type_info(), shape + 2); // 2 = ": "
 
     type_specifier
         .to_owned()